Refactor NavigationTransition to TransitionAction.
