+++ Example/SwiftUICoordinatorExample/SwiftUICoordinatorExample/Coordinators/Shapes/ShapesCoordinator.swift
@@ -55,7 +55,7 @@ class ShapesCoordinator: NSObject, Coordinator, Navigator {
     // MARK: - Private methods
     
     private func setup() {
-        navigationController.register(FadeAnimator())
+        navigationController.register(FadeTransition())
     }
 
     private func makeSimpleShapesCoordinator() -> SimpleShapesCoordinator {
@@ -86,35 +86,6 @@ extension ShapesCoordinator: RouterViewFactory {
             CustomShapesView<CustomShapesCoordinator>()
         case .featuredShape:
             EmptyView()
-        }
-    }
-}
-
-class FadeAnimator: NSObject, Transition {
-    func isEligible(from fromRoute: NavigationRoute, to toRoute: NavigationRoute) -> Bool {
-        return (fromRoute as? CustomShapesRoute == .customShapes && toRoute as? CustomShapesRoute == .star) ||
-        (fromRoute as? CustomShapesRoute == .customShapes && toRoute as? CustomShapesRoute == .tower)
-    }
-    
-    func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
-        return 0.3 // Set the duration of the fade animation
-    }
-    
-    func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
-        guard let toView = transitionContext.view(forKey: .to) else {
-            transitionContext.completeTransition(false)
-            return
-        }
-        
-        let containerView = transitionContext.containerView
-        toView.alpha = 0.0
-        
-        containerView.addSubview(toView)
-        
-        UIView.animate(withDuration: transitionDuration(using: transitionContext), animations: {
-            toView.alpha = 1.0
-        }) { _ in
-            transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
         }
     }
 }
