+++ Example/SwiftUICoordinatorExample/SwiftUICoordinatorExample/Transitions/FadeTransition.swift
@@ -1,0 +1,38 @@
+//
+//  FadeTransition.swift
+//  SwiftUICoordinatorExample
+//
+//  Created by Erik Drobne on 15/05/2023.
+//
+
+import SwiftUI
+import SwiftUICoordinator
+
+class FadeTransition: NSObject, Transition {
+    func isEligible(from fromRoute: NavigationRoute, to toRoute: NavigationRoute) -> Bool {
+        return (fromRoute as? CustomShapesRoute == .customShapes && toRoute as? CustomShapesRoute == .star) ||
+        (fromRoute as? CustomShapesRoute == .customShapes && toRoute as? CustomShapesRoute == .tower)
+    }
+    
+    func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
+        return 0.3 // Set the duration of the fade animation
+    }
+    
+    func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
+        guard let toView = transitionContext.view(forKey: .to) else {
+            transitionContext.completeTransition(false)
+            return
+        }
+        
+        let containerView = transitionContext.containerView
+        toView.alpha = 0.0
+        
+        containerView.addSubview(toView)
+        
+        UIView.animate(withDuration: transitionDuration(using: transitionContext), animations: {
+            toView.alpha = 1.0
+        }) { _ in
+            transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
+        }
+    }
+}
