+++ Sources/SwiftUICoordinator/Navigator/NavigatorError.swift
